Reorganize repo to improve imports
