+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "corpus-query-language"
-version = "0.0.2"
+version = "0.0.4"
 authors = [
   { name="Matthias Gille Levenson", email="matthias.gille-levenson@ens-lyon.fr" },
 ]
@@ -20,6 +20,9 @@ dependencies = [
 [build-system]
 requires = ["setuptools", "setuptools-scm"]
 build-backend = "setuptools.build_meta"
+
+[tool.setuptools.packages.find]
+where = ["src"]
 
 [project.urls]
 Homepage = "https://github.com/matgille/CQL"
