+++ src/corpus_query_language/__init__.py
@@ -1,0 +1,5 @@
+import corpus_query_language.core.core as core
+import corpus_query_language.engine.engine as engine
+import corpus_query_language.utils.utils as utils
+import corpus_query_language.language as language
+__all__ = ["core", "engine", "language", "utils"]
