+++ tests/tests.py
@@ -1,8 +1,7 @@
 import ast
 import sys
 sys.path.append('src/')
-import corpus_query_language.core.core as core
-import corpus_query_language.functions.functions as functions
+import corpus_query_language as CQL
 import unittest
 
 def import_test_queries(path):
@@ -24,14 +23,14 @@ class TestFunctions(unittest.TestCase):
 					  "pos": "NCMS000",
 					  "morph": None,
 					  "word": "asnos"}
-		self.assertEqual(functions.simple_match(query, test_token), True, "Something is wrong"
+		self.assertEqual(CQL.utils.simple_match(query, test_token), True, "Something is wrong"
 																		  "with function `test_simple_match`")
 
 class TestQueries(unittest.TestCase):
 	def test_findall_queries(self):
-		self.corpus = functions.import_corpus("tests/test_data/test_corpus.json")
+		self.corpus = CQL.utils.import_corpus("tests/test_data/test_corpus.json")
 		self.queries = import_test_queries("tests/queries_findall.txt")
-		self.MyEngine = core.CQLEngine()
+		self.MyEngine = CQL.core.CQLEngine()
 		for query, GT in self.queries:
 			GT = ast.literal_eval(GT)
 			with self.subTest(query=query, GT=GT):
@@ -40,7 +39,7 @@ class TestQueries(unittest.TestCase):
 
 	def test_match_queries(self):
 		self.queries = import_match_queries("tests/queries_match.txt")
-		self.MyEngine = core.CQLEngine()
+		self.MyEngine = CQL.core.CQLEngine()
 		for idx, (nodes, query, GT) in enumerate(self.queries):
 			with self.subTest(query=query, GT=GT):
 				GT = True if GT == "True" else False
